I added a Click event that for all operations. I made the clear buttons make the field zero.
The label on top partially works.

Have not done anything for the radio buttons yet, besides research.

Im in the midst of optimizing the code for each of the number buttons, by adding the same click event to each number button like for the operation buttons. This will shorten the code.